Add debian/watch file.
A watch file is useful to monitor when upstream uploads new releases.
It is used by running `uscan' in the directory containing the debian/
subdirectory.

This file mangles usual pre-releases suffixes such as dev, rc, alpha,
etc. to make sure that the final version (without any suffix) is
considered newer. It also supports various files extensions in case
next releases are using a different compression scheme.

See:
- http://www.debian.org/doc/manuals/maint-guide/dother.en.html#watch
- https://wiki.debian.org/debian/watch/

The lack of this file is triggering a Lintian pedantic notice:
http://lintian.debian.org/tags/debian-watch-file-is-missing.html